+++ utils/index.js
@@ -31,6 +31,7 @@ module.exports = (...args) => {
       library.utils["powershell"] = require("./powershell")(params, obj);
       library.utils["intercomm"] = require("./intercomm")(params, obj);
       library.utils["sqlitesession"] = require("./sqlitesession");
+      library.utils["csse"] = require("./csse.js")(params, obj);
       library.utils["smfetch"] = require("./smfetch")(params, obj);
       library.utils["startupinit"] = require("./startupinit")(params, obj);
       library.utils["excluded"] = [