+++ Changelog.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Move the startupinit module from atomic/atom to utils. Completed on 2026-01-01
+- Completelly migration atom/smfetch and atom/csse to utils. Wait for test.
 
 ### Deprecated